Improve performance

+++ lib/aphrael/directory.rb
@@ -21,20 +21,30 @@ class Aphrael::Directory < Aphrael::Resource
     @children
   end
 
-  def create_index
+  def images length=nil
+    create_index length
+    @images
+  end
+
+  def create_index length=nil
     unless defined?(@children)
       @children = []
+      @images = []
       Dir.entries(real_path).sort.each do |file|
+        if length && @images.size >= length
+          break
+        end
         next if file.start_with?('.')
         real_file_path = File.join(real_path, file)
         if File.directory?(real_file_path)
           _path = [path, file]
           _path.delete('')
           @children << self.class.get(index, File.join(*_path))
+        elsif File.file?(real_file_path) && Aphrael::Image::SUPPORTED_FORMATS.include?(File.extname(real_file_path))
+          @images << Aphrael::Image.get(index, File.join(path, file))
         end
       end
     end
-    @children
   end
 
   def to_h
